requirements.txt
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
scipy>=1.10.0
sentence-transformers>=2.2.0
tqdm>=4.65.0
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
openpyxl>=3.1.0
python-docx>=1.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
sqlite-vss>=0.1.0
neo4j==5.14.1
